chore: move `getIconForColumnType` to transform
